Cast the initialized image shape to integer
Fix the error 'numpy.float64' object cannot be interpreted as an integer

+++ display_network.py
@@ -31,7 +31,8 @@ def display_network(A, m = -1, n = -1):
         m = np.ceil(col / n)
         
     # Initialization of the image
-    image = np.ones(shape=(buf + m * (sz + buf), buf + n * (sz + buf))) # Using np.ones to make white border btw digits
+    #print(buf + m * (sz + buf), buf + n * (sz + buf))
+    image = np.ones(shape=(int(buf + m * (sz + buf)), int(buf + n * (sz + buf)))) # Using np.ones to make white border btw digits
 
     if not opt_graycolor:
         image *= 0.1
